Separate export button in a separate component
